+++ R/helper_exec.R
@@ -28,12 +28,16 @@ future_map = function(n, FUN, ..., MoreArgs = list()) {
       mirai::daemons_set(.compute = getOption("mlr3.mirai_parallelization", "mlr3_parallelization"))
   ) {
     lg$debug("Running resample() via mirai with %i iterations", n)
-    mirai::collect_mirai(mirai::mirai_map(
-      data.table(...),
-      FUN,
-      .args = c(MoreArgs, list(is_sequential = FALSE)),
-      .compute = getOption("mlr3.mirai_parallelization", "mlr3_parallelization")
-    ))
+    # .stop re-raises errors from the workers instead of returning them as `miraiError` values
+    mirai::collect_mirai(
+      mirai::mirai_map(
+        data.table(...),
+        FUN,
+        .args = c(MoreArgs, list(is_sequential = FALSE)),
+        .compute = getOption("mlr3.mirai_parallelization", "mlr3_parallelization")
+      ),
+      ".stop"
+    )
   } else {
     is_sequential = inherits(plan(), "sequential")
     scheduling = if (!is_sequential && isTRUE(getOption("mlr3.exec_random", TRUE))) {